# docs/swagger.yaml
basePath: /
info:
  contact:
    email: geek@openmind.np
    name: Above&Beyond
  description: API Doc for Suglider Auth
  license:
    name: MIT
    url: https://github.com/tony/suglider-auth/blob/master/LICENSE
  termsOfService: http://swagger.io/terms/
  title: Suglider-Auth API Doc
  version: "1.0"
paths:
...
  /api/v1/user/forgot-password:
    get:
      consumes:
      - application/json
      description: Send a mail to reset password
      parameters:
      - description: Email
        in: query
        name: mail
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Send Password Reset Email
      tags:
      - users
...
  /api/v1/user/reset-password:
    post:
      consumes:
      - application/json
      description: reset password for user
      parameters:
      - description: New Password
        in: formData
        name: password
        type: string
      - description: Email
        in: query
        name: mail
        type: string
      - description: Reset ID
        in: query
        name: reset-id
        type: string
      - description: Reset Code
        in: query
        name: reset-code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Reset Password
      tags:
      - users
...
  /api/v1/user/verify-mail:
    post:
      consumes:
      - application/json
      description: verify email address for user
      parameters:
      - description: Email
        in: query
        name: mail
        type: string
      - description: Verify ID
        in: query
        name: verify-id
        type: string
      - description: Verify Code
        in: query
        name: verify-code
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Verify Email Address
      tags:
      - users
  /api/v1/user/verify-mail/resend:
    get:
      consumes:
      - application/json
      description: Resend a mail to verify email address for user
      parameters:
      - description: Username
        in: query
        name: username
        type: string
      - description: Email
        in: query
        name: mail
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
      summary: Resend Verify Email
      tags:
      - users
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-KEY
    type: apiKey
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"